requirements.txt
asgiref==3.5.0
django-cors-headers==3.11.0
django-filter==21.1
Django==3.2.12
djangorestframework==3.13.1
importlib-metadata==4.11.1
Markdown==3.3.6
pip==21.1.2
pytz==2021.3
setuptools==57.0.0
sqlparse==0.4.2
wheel==0.36.2
zipp==3.7.0
